requirements.txt
# ─────────────────────────────────────────────
# Configurable Workflow Decision Platform
# Python Dependencies
# ─────────────────────────────────────────────

# Web Framework
fastapi==0.115.5
uvicorn[standard]==0.32.1

# Database (async SQLAlchemy + SQLite + PostgreSQL drivers)
sqlalchemy[asyncio]==2.0.36
aiosqlite==0.20.0
asyncpg==0.30.0           # PostgreSQL async driver (for production)

# Configuration
pyyaml==6.0.2

# Validation
pydantic==2.10.3
pydantic-settings==2.6.1

# HTTP Client (for testing)
httpx==0.28.1

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0

# Utilities
python-multipart==0.0.20  # For form data support
